[#23] 패키지 구조 변경  - move to package com.stos -> com.stos.concert
