+++ hall-service/src/main/java/com/stos/concert/HallServiceApplication.java
@@ -1,4 +1,4 @@
-package reservation;
+package com.stos.concert;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
