+++ hall-service/src/main/java/com/stos/concert/hall/domain/Hall.java
@@ -1,0 +1,15 @@
+package com.stos.concert.hall.domain;
+
+import java.util.List;
+
+public class Hall {
+	private final Long id;
+	private final String name;
+	private final List<Seat> seats;
+
+	public Hall(Long id, String name, List<Seat> seats) {
+		this.id = id;
+		this.name = name;
+		this.seats = seats;
+	}
+}
