--- hall-service/src/main/java/com/stos/schedule/entity/ScheduledSeat.java
@@ -1,22 +1,0 @@
-package com.stos.schedule.entity;
-
-public class ScheduledSeat {
-    private final Long seatId;
-    private final SeatStatus seatStatus;
-
-    public ScheduledSeat(Long seatId, SeatStatus seatStatus) {
-        this.seatId = seatId;
-        this.seatStatus = seatStatus;
-    }
-
-    enum SeatStatus {
-        FREE,
-        TEMPORARY_OCCUPIED,
-        RESERVED,
-        CLOSED,
-    }
-
-    public boolean isReservable(){
-        return this.seatStatus.equals(SeatStatus.FREE);
-    }
-}
